Draw parameters with fixed effects

Draws parameters directly from the distributions specified by model_pars for the antibody model with fixed effects. This function ensures that all individuals have the same parameters.

Usage

draw_parameters_fixed_fx(
  i,
  t,
  x,
  b,
  demography,
  biomarker_states,
  model_pars,
  ...
)

Arguments

i

individual

t

time

x

exposure

b

biomarker

demography

demography information

biomarker_states

an array of true biomarker quantities for all individuals across all time steps and biomarkers

model_pars

tibble of biomarker (antibody) kinetics parameters with variables: 1) exposure_id: numeric exposure ID; 2) biomarker_id: numeric biomarker ID; 3) name: the character name of the parameter; 4) mean: numeric mean of this parameter distribution; 5) sd: the numeric standard deviation of the parameter distribution; 6) distribution: character description of the parameter distribution type (e.g., log-normal, normal)

...

Additional arguments

Value

A tibble with the simulated parameters for this exposure event

Examples

draw_parameters_fixed_fx(1,1,1,1,example_demography, example_biomarker_states, 
example_model_pars_numeric)
#> # A tibble: 2 × 7
#>       i     t     x     b name   value realized_value
#>   <dbl> <dbl> <dbl> <dbl> <chr>  <dbl>          <dbl>
#> 1     1     1     1     1 boost 4              4     
#> 2     1     1     1     1 wane  0.0033         0.0033