Example immune histories (wide)

Example immune history runserosim output. This example data set output contains immune history for individuals at all time steps for each exposure ID. This is identical to example_immune_histories, but is converted into a 3D array giving the format expected by most antibody model functions.

data(example_immune_histories_wide)

Format

A 3D array with dimensions corresponding to 1) individual; 2) time; 3) exposure ID, and values giving the exposure state

See also