Example numeric model parameters (model_pars)

Example numeric model parameters (model_pars) required as an input for runserosim. This example model_pars is for a simulation two different exposure events both containing the same biomarker. model_pars argument is responsible for storing parameter information needed for the antibody model, observation model and immunity model. In this example, model_pars has parameters for a monophasic antibody model and for an observational model with noise.

data(example_model_pars_numeric)

Format

A data frame with 5 rows and 6 variables:

exposure_id

numeric values for each exposure type present in example biomarker_map

biomarker_id

numeric values for each biomarker type present within each exposure type present in example biomarker_map

name

names of model parameters

mean

numeric values for the true paramter means

sd

numeric values for the true paramter standard deviation

distribution

distribution type from which the draw_paramaters function will be sampling a parameter

See also