Fixed probability of exposure

Generic wrapper function to allow the probability of exposure for each time point to be drawn directly from the foe_pars array.

exposure_model_fixed(i, t, x, g, foe_pars, demography, ...)

Arguments

i

individual

t

time

x

exposure

g

group

foe_pars

A 3D array providing the probability of exposure for each exposure ID, group and time.

demography

A tibble of relevant demographic information for each individual in the simulation.

...

Additional arguments

Value

A probability of exposure from the foe_pars array

Examples

times <- seq(0,365,by=1)
n_groups <- 1
n_exposures <- 1
foe_pars <- array(0.01, dim=c(n_groups,length(times),n_exposures))
exposure_model_fixed(1,1,1,1,foe_pars,NULL) 
#> [1] 0.01