Simulate random birth times

This function simulates uniformly distributed birth times for a specified number of individuals from a provided vector

simulate_birth_times(N, times, age_min = 0)

Arguments

N
The number of individuals in the simulation

times
A vector of each time step in the simulation

age_min
A number matching the time resolution of times giving the youngest age possible by the end of the simulation; defaults to 0 which means individuals can be born up until the penultimate time step

Value

A vector of simulated birth times for each individual

See also

Examples

## Simulate random birth times for 500 individuals over 100 time steps and ensures 
## that all individuals are above 9 time steps old by the last time step
simulate_birth_times(500, 1:100, age_min=9) 
#>   [1] 81 30 34 64 59 36 23 76 50 26 57 78 89 33 80 60 12  5 60 69 30 15 86 12 68
#>  [26] 40 64 35  2 68 64 26 40 13 65 27 60 31 88 52 54 29  3 86 41 76 85  8  9 22
#>  [51]  5 24 89 33 17 87 58 82 62  6 57 71  2 31 61 21 88 81 51 31 22 72 44 20 75
#>  [76] 84 61  6 30 15 18 42 70 62 19 49 37 41 68 23 76 21  3 63 78 28 88 29 20 61
#> [101] 22 47 52  9 24 29 62 24 66 69 87 51 33 58 59 68 56 59 70 35 14 23 46 23 40
#> [126] 20 70 50 71 80 70 22 26 38 51  2 11 20 31 71  2 87 60 61 54 36 48 29 69 25
#> [151] 83 67 72 29 78 66 35 70 24 20 19 15 43 90 55 60 69 39 35  8 32 30 89 22 45
#> [176] 63 21 84 70 59 79 52 37 38  6 38 61 79 45 21 64 87 68 86 33 67 46 37 80 43
#> [201] 33 82 23 26  2 51 85 34  2 74 56 38 30 77 81 27 10 54 77 52 28 77 13 72 49
#> [226] 54 23 57 74  5 67 81 48 46 28 75 81 49 51 89 48 71 54 35 87 50 76 66 53 89
#> [251] 34 63  8 28 90 64  2 26 89 56 54 16 35 50 27 69 26 10 15 72 25 34 80 69 37
#> [276] 12 22 52 74 32 65 43 65 82 81  5 79 28 18 58 90 65 31 82 24 69  6 34 88  8
#> [301]  6 68 55  4 34 34 51 43 50  9 19  2 79 16 42 15 27 48 74 80 50 31 81 24 88
#> [326] 66 55 83 14 13 75 20 72 46 67 69 58 51 30 69 28 80 30  7  2 15 35 19 25 58
#> [351] 83 37 28 66 32 26 41 86 16 89 74 31 12 44 81 55 88 90 78 23 78  4 42 20 24
#> [376] 70 10 60  9 78 50 37 81 88 48 34 59 25  9 68 11 70 73 24 30 55 33 61 55  3
#> [401] 14 73 25 78 89 81 59 64 50 60 29 72 16 16 54  9 55 65 82 13 42 82 90 26 13
#> [426] 19 43 38 62 54 47 70 70  1 28 30 53 20 44 38 72 56  5  7  8 18 64 61 75 69
#> [451] 43 62 54 20 87 33 14 48 75 58 56 69 48 86 57 55 77 10 24 15 40 75 58  2 44
#> [476] 28 66 65  6 10  3 20 38 53 17 36  4 24 14 59 13 28  5 44 66 43 57 85 84 63