Simulate removal times for individuals

This function simulates uniformly distributed removal times for a specified number of individuals from the provided times vector. This might represent e.g., death or study attrition.

simulate_removal_times(
  N,
  times,
  birth_times,
  removal_min = 0,
  removal_max = max(times),
  prob_removal = 0
)

Arguments

N

The number of individuals in the simulation

times

A vector of each time step in the simulation

birth_times

A vector of birth times for each individual; defaults to NULL; if birth_times is not specified then the function will simulate uniformly distributed birth times for each individual from the times vector

removal_min

The minimum age at which an individual can be removed from the population. Defaults to 0

removal_max

The maximum age at which an individual can be removed from the population. Defaults to max(times)

prob_removal

The probability that an individual will be removed from the population during the simulation, representing e.g., death or study attrition. If set to NA, then removal time will be max(times)+1

Value

A vector of all individual's removal times is returned. NA represents no removal

See also

Examples

## Simulate random removal times for all individuals; Individuals have a 0.4 probability 
## of being removed at some time after they are 10 time steps old and before 
## they are 99 time steps old 
birth_times<-simulate_birth_times(500, 1:100, age_min=9) 
simulate_removal_times(500,1:100,birth_times, removal_min=10,removal_max=99, prob_removal=0.4)
#>   [1] 101 101 101 101 101 101 101 101  75  63  39 101 101 101 101 101 101 101
#>  [19]  78 101 101  87 101 101 101  96 101  36 101  82 101 101 101 101 101  67
#>  [37]  19 101 101 101 101  49 101 101 101 101  99 101  23  88 101 101 101  44
#>  [55] 101  83 101  95  11 101 101  96  68  48  64  89  64  98  55 101 101  87
#>  [73]  94  95 101  51  53  67  98 101  94  63 101 101 101 101 101 101 101  47
#>  [91] 101 101  66 101 101  88 101  98 101  58  75  33  24  23 101  90  92  76
#> [109] 101  98 101  78 101 101  86 101  86 101  86 101 101 101 101 101  97 101
#> [127]  94 101 101 101  60 101  47  82  71  38 101 101  66  86  83 101 101 101
#> [145] 101  98 101 101 101  53  96  87  92 101 101  69  37  65  59  95  57  71
#> [163] 101  20  48 101 101 101  56 101 101 101 101 101 101  46 101  79 101  80
#> [181]  86 101 101 101 101  48 101  85 101  88 101 101 101  73  82 101 101 101
#> [199] 101  96 101 101 101 101  82  79 101 101  84  54  95 101  83 101  93 101
#> [217]  26  77 101 101  27 101 101 101  37  51  76 101  70 101 101 101  97 101
#> [235]  97  61 101 101 101  97 101 101 101  72 101 101  89 101 101 101 101  14
#> [253] 101  81  56 101  77 101 101  93 101 101  87 101 101 101 101  70 101  45
#> [271] 101 101 101 101 101  83  69  98 101  73 101 101 101 101 101  22 101 101
#> [289]  48  83 101 101  89  85 101  77  43 101 101  29  95 101  94 101  58 101
#> [307]  72  65 101 101  61 101  84  87 101 101  83 101  55  92  57 101 101 101
#> [325] 101  74 101  70 101 101 101 101  37 101  64  78 101  65  87 101 101  81
#> [343]  80  95 101  68 101 101  69 101  45  76 101  60 101  63  67 101 101 101
#> [361]  66  75 101 101  98 101  86  70 101 101 101 101  92 101 101  87 101 101
#> [379]  49  99  65  95 101 101  61 101 101  60 101 101 101 101 101 101 101 101
#> [397]  74 101 101 101  35 101  37  41 101  89 101 101 101 101 101  92 101  93
#> [415]  95  90  94 101  54 101 101  68 101  69  90  70 101  42 101 101 101  51
#> [433] 101 101  92  24  88  66  99  36 101 101 101  79 101  72  93 101 101 101
#> [451]  80 101  56 101 101 101  83 101 101  86  65  38  95 101  91 101 101 101
#> [469] 101  80  80  38 101 101  75  80  86  98  94 101 101  33  25 101  73 101
#> [487]  49  76  46  46 101 101 101 101  77 101 101 101  42  54