Observation model for continuous assays with no detection limits and added noise

This observation model observes the latent biomarker quantities given a continuous assay with added noise. The added noise represents assay variability and is done by sampling from a distribution with the latent biomarker quantity as the mean and the measurement error as the standard deviation. The observation standard deviation and distribution are defined within model_pars as the obs_sd parameter. The user can also use the optional sensitivity and specificity arguments to account for assay sensitivity and specificity. False negatives are simulated by setting an observed quantity to 0 with probability sensitivity. False positives are simulated by drawing a random quantity from the bounded range for a true 0 biomarker quantity with probability 1-specificity.

Usage

observation_model_continuous_noise(
  biomarker_states,
  model_pars,
  sensitivity = 1,
  specificity = 1,
  ...
)

Arguments

biomarker_states

tibble containing true biomarker quantities for all individuals across all time steps and biomarkers. Variables should include: 1) i: the individual ID; 2) t: the time period; 3) b: the biomarker ID; 4) value: the latent biomarker quantity for the given i, t and b

model_pars

a tibble containing information for all parameters needed to simulate the observation process. This should usually contain: 1) exposure_id: numeric exposure ID; 2) biomarker_id: numeric biomarker ID; 3) name: the character name of the parameter; 4) mean: numeric mean of this parameter distribution; 5) sd: the numeric standard deviation of the parameter distribution

sensitivity

number between 0 and 1 to describe the assay's sensitivity; defaults to 1

specificity

number between 0 and 1 to describe the assay's specificity; defaults to 1

...

Additional arguments

Value

biomarker_states is returned with a new column, observed, for observed biomarker quantities

See also

Examples

observation_model_continuous_noise(example_biomarker_states, example_model_pars_numeric, 0.95,0.99)
#>          i   t b    value    observed
#>     1:   1   1 1  0.00000  0.20158825
#>     2:   1   2 1  0.00000 -0.01596533
#>     3:   1   3 1  0.00000 -0.05218665
#>     4:   1   4 1  0.00000 -0.29147948
#>     5:   1   5 1  0.00000 -0.03652252
#>    ---                               
#> 11996: 100 116 1 16.26414 16.56695792
#> 11997: 100 117 1 16.20296 16.21081648
#> 11998: 100 118 1 16.14178 15.96934211
#> 11999: 100 119 1 16.08060 16.21108495
#> 12000: 100 120 1 16.01942 16.24817963