Simulate random birth times

This function simulates uniformly distributed birth times for a specified number of individuals from a provided vector

Usage

simulate_birth_times(N, times, age_min = 0)

Arguments

N

The number of individuals in the simulation

times

A vector of each time step in the simulation

age_min

A number matching the time resolution of times giving the youngest age possible by the end of the simulation; defaults to 0 which means individuals can be born up until the penultimate time step

Value

A vector of simulated birth times for each individual

See also

Examples

## Simulate random birth times for 500 individuals over 100 time steps and ensures 
## that all individuals are above 9 time steps old by the last time step
simulate_birth_times(500, 1:100, age_min=9) 
#>   [1] 33 29 42 76 22 85 52 37 41 39 16 29 67 86 90 28 61  1 14 71 69 45 11 84 78
#>  [26] 83 43 10 53 20  1 51  7 13 56 90 82  8 70 79 72 79 29 85 57 45 54 86 87 45
#>  [51] 28 15 17 11 10 25 57 15 27 70 34 19 47 69 33 76 57 78  9 85 10 69 61 18 24
#>  [76] 60 12 81 55 87 48 51  8 72 43 44 88 12 58 80 64 57 80 47 55 28 47 45 50 14
#> [101] 47 90 18 82 20 32 56 73 10  9 65 35 57 89 68 70 85 54  1 56 36 18  1 35  3
#> [126] 80 13 51 76 90 69 74 61  4  6 74 49 63 73 54 77 79 36  7 53 83 29 64 85  1
#> [151] 80 13 89 42 46 73 52 37 21  6 22  4 86 55 36 75 42 38 84 27 58 63 23  3 46
#> [176] 26 63 50  2 18  3 38 32 13 79 52 52  1 27 40 31 18 57 50  1 27 19  2 69 55
#> [201] 58 68  5 23  3 84 40 69 81 36 83 69 14 58 89 20 87 78 85 35 62 68 57 76 43
#> [226] 23  1 13 46 26 18 31 17 75 61 63 82 42 26 70 25 79 38 35 38 87 44  5 25 88
#> [251] 82 32 73 41  1 88 64 13 79 18  4 83 83 53 88 64 47 14 78 76 36 63 32 30 39
#> [276] 56 84 75  1 67 75 34  2  4 76  7 59 16 19 13 57 56 56 40 40 29 81  8 55 74
#> [301] 41 75 32  3 74 45 17 89 57 71 27 76 62 18 15 13  9  2  2 80 23 44 63 60 72
#> [326] 13 72 23 49 90 19 25 66 68 22 11 61 86 71 48 75 74 44 81 34  1 36 56 54 24
#> [351] 44 50 28 34 47 53 70 11  8 64 26  5 36  3 12 38 17 19 78 90 27 29 26 26  8
#> [376] 75  2 71 67 89 84 32  1 61 67 17 21 90 44 69 75 70 10 29 43 83  6 51 71 57
#> [401] 25 55 60 42 83 30 34 17 66 19 55 25 35 16 67 11 45 83  8 41 24 63  7 75 33
#> [426] 79 46 86 79 40  9 25 75 42 58 29 50 63 68 49 53 34 27 60 14 81 77 87 10 77
#> [451] 61 16 75 73 64 65 46 42 79 56 47 35 85 57 28 61 36 11 17 74 62 54 60 19 22
#> [476] 13 27 82 28 68 48 42 66 20 79 75 13 23 59 85 71 86  9 17 43  2 27 69  1 45