Simulate random birth times

This function simulates uniformly distributed birth times for a specified number of individuals from a provided vector

Usage

simulate_birth_times(N, times, age_min = 0)

Arguments

N

The number of individuals in the simulation

times

A vector of each time step in the simulation

age_min

A number matching the time resolution of times giving the youngest age possible by the end of the simulation; defaults to 0 which means individuals can be born up until the penultimate time step

Value

A vector of simulated birth times for each individual

See also

Examples

## Simulate random birth times for 500 individuals over 100 time steps and ensures 
## that all individuals are above 9 time steps old by the last time step
simulate_birth_times(500, 1:100, age_min=9) 
#>   [1] 55 60 38 79  9 12 73 27  2 10 90 12 78  9 17 67 83 14 69 84 70 87 41 42 22
#>  [26] 24 71 62 31 48 40 51 57 39  1 86 28 80 35 11  8 49 29 44 35 11 55 11 84 23
#>  [51] 33 67  3 27 54 17 61 22  8 46 52 29 14 17 22 45 24  6 55 75  4  1 38 31 78
#>  [76] 13 36 58 48  4 69 39 69 22 56 58 34 40 87 87 40 25 89  8  7 90 75 45 34 84
#> [101] 13 17 80 71 89 22 41 25 32 13 50 37 30 26 23 49 86 45  1 34 67  2 38 20 46
#> [126] 26 76 48 77 47 83 66 12 55 25 87 84 20 58  2 10 11 84 53 87 61 32 67 27 28
#> [151] 44 53 78 27 52 26 74 60 26 20 22 86 70 69 29 38 32 31 40 15 57 69 82 43 18
#> [176] 53 22 90 86 38 25 36 81 85 59 87 36 34 34 43 62 59 21 63 40 77 23 30  5 42
#> [201] 23 70 11  1 39 25 79 30 25 85 39  9 41 89 64 63  8 89 41 44 88 77 34  3 39
#> [226] 74 15 69 87 31 80 41 77 24  2 82 67 25 28 48 11 17 60 88  8 10 22 29 35 41
#> [251] 10 37 10 79 84 87  9 25 20 12 62 90 12 90  6 90 72 64 19 59 89 12 57  3  8
#> [276] 36 18 33 74  5 10  5 40 62  1 13 80 56 44 44 76 85 55 42 80 51 24 21 54 69
#> [301] 22 34 30 60 14 52 59 16 86 19 29 55 75 83  6  4 46 53 85 10 12 24 82 42 61
#> [326] 18 40 70 36  4 63 83 63 10 18 12 37 82 87 62 18 66  4 49  1 84 15 59 71 22
#> [351] 14 10 15 56 27 39 40 66 64  9 44 66 22 11 81 33 42 30  2 86 54  4  3 30 24
#> [376] 45  9 29 22 85 82 12 75 56 30 20 23  3 61 76 43 68 83 60 69 52 41 24 62 57
#> [401] 63 30 80 10 50 63 18 29 85  9 59 13  2 54 14 87 89 79 17  1 33 79 25 50 34
#> [426] 21 33 61 71 72 64 61 33 24 80 88  6 42 37 70 32 22  7 45 39 46 33 75 17 37
#> [451] 70 30 90 42 25 19 48 53 50 32 20 57 44  9 16 85 20 80 73  9 43 76 32 71  3
#> [476] 61 80 84 19 53 21 51 19 23 84 80 13 60 46 57 34 35 53 62 48  6 57 69  1 83