Simulate random birth times

This function simulates uniformly distributed birth times for a specified number of individuals from a provided vector

simulate_birth_times(N, times, age_min = 0)

Arguments

N

The number of individuals in the simulation

times

A vector of each time step in the simulation

age_min

A number matching the time resolution of times giving the youngest age possible by the end of the simulation; defaults to 0 which means individuals can be born up until the penultimate time step

Value

A vector of simulated birth times for each individual

See also

Examples

## Simulate random birth times for 500 individuals over 100 time steps and ensures 
## that all individuals are above 9 time steps old by the last time step
simulate_birth_times(500, 1:100, age_min=9) 
#>   [1] 32 62 84 46 78 25 83 35 56 80 65 17 60 18 23 53 83 24 32 30 46 32 83 42 64
#>  [26] 84 83 51 52 86 80 60 44 28 19 82 28 89 18 52 75 83 56 43 16 89 46 90 77  4
#>  [51] 13 21 63 40 35 80 87 56 15 42 72 37 67 34 47 89 44 80  8 30 35 74 80 81  1
#>  [76] 31 64 85 45 11 34 73 81 26 64 10 31 70 89 23  3  6 39 33 52 46 29 45 15 31
#> [101] 51 89 69 33 69 16 13 26  9  1 16 75 78  4 45 85 50 15 12 85 72  6 53 15 84
#> [126] 50 72 59 29  9  2 69 73 60 12 88 38 40 81 57 36 88 71  1 59 54 56 20 18 69
#> [151] 52 67 59 88 72 86 85 85 50 18 33 53 57 54 40 23 30 38 57 72 27 80 87 49 32
#> [176] 12 57 20 22 73 74  4 59 64 21 15 11 43 81 36 89  3 84 43 44  4 81 85 74 52
#> [201] 29 48 33 10 84 52 82 45 23 14 71 62 11 51  4 62 81 46 49 86 43 82 53 56 53
#> [226] 75 53 15 26 53 78 85  1 87 77 82 60 48 46 73 30 23 80 66 37 34 47 54  5 29
#> [251] 63 38 53 64 48 80 12 76 17 33 85 55 69 75 49 69 39 71 72 63 86 70 46 54 73
#> [276] 75 26 26 86 57 83 60 59 35 67  1 30 14 43 46 14 53 68 44 85 45 70 74  4 48
#> [301] 84 59 20 78 66 29 16 87 70 85 48 11 27 76 23 49 67 29 47 75 24 15  2 24 42
#> [326] 11 54 19 83 58 23 12 22 16 36 88 22 79 87 28 66 59 62  4 24 75 24 74 71 34
#> [351] 49 73 18 47 88 29 84 57 32 10 15 71 27 53 47 43 24  2 16 42 68 69 74 28 56
#> [376]  4 45 29 79 32 69 60 37 29 16 62 86 87 71 57 76 58 42 65 78 70 31 11 87  4
#> [401] 63 18 38 87 77 67 87 62 50  4  4 40 32 67 47 39 41 78 81 86 57 51 71 67 24
#> [426] 36 87 56  8 80 63 50 83 58 17 79 38 43 16 80 87 63 65 21 13 60 27  1 63 43
#> [451] 76 74 24 52 48 15 75 81 77 40 60 71 70 10 89  3 55 54 55 19 22 75 60 39 44
#> [476] 79 74 30 13 69 72  4 68  3 40 78 11 18 89 70 85 71 60 22 67 64 72  1 69 84