Simulate random birth times

This function simulates uniformly distributed birth times for a specified number of individuals from a provided vector

Usage

simulate_birth_times(N, times, age_min = 0)

Arguments

N

The number of individuals in the simulation

times

A vector of each time step in the simulation

age_min

A number matching the time resolution of times giving the youngest age possible by the end of the simulation; defaults to 0 which means individuals can be born up until the penultimate time step

Value

A vector of simulated birth times for each individual

See also

Examples

## Simulate random birth times for 500 individuals over 100 time steps and ensures 
## that all individuals are above 9 time steps old by the last time step
simulate_birth_times(500, 1:100, age_min=9) 
#>   [1]  9 13  9 73 74 11 61 62 37 84 69 13 32  3  9 14 58 50 32 36 24 63 48 57 29
#>  [26] 39 34 66  7 32 90 12  9 70 59 60 27 23 56 19 60 62 55 25 23 18 77 89 24 47
#>  [51] 44 22 82 36 25 70 40  4 54 27 67 35 44 81 66  1 19 35 74 83 35 54 74 74 18
#>  [76] 65 12 25 65 23 13 66  9 42 61 79 63 11 68 31 29  8 42 22 42 81 70 42 37 83
#> [101] 39 68 50 24  9 25 72 75 71 67 59 72 10 62 37 18 45 60 53 31 50  8 48 51 45
#> [126] 70 36 15 79 78 84 89 37 63  7 31 50 79  3 87 55 13 29 65 85 61 29 12 90 50
#> [151] 41 82 73 83  6 60  2 18 53 12 70 86 19 64 10 49 71 53 62 14 46 23 54 86 21
#> [176] 63 61 74 30  5  1  6 46 84 13 65 90 20 25 63 75 43 44 79 54 71 29  8 26 21
#> [201] 84 78 85 79 43 13 47 56 74 90 60 80 52 77 63 73 64 64 73 72 84 84 18 46 34
#> [226] 57 20 73 27 11 61 13 21 51 67 51 78 16 25 63 16 47 33  6 87 70 55 14 52 61
#> [251] 69 44 54 75 58 65 39 12 57  2 58 83 54 88 43 39 73 24 48 22 60 74 64 15 66
#> [276]  2 45 40 44 65 14 13 28 13 31 79  6 69 33 58 21 24 75 70 41 56 63  2 27 77
#> [301] 21 61 84 33 50 45 51 23 13  1  4 20 90 33 44 78 29 65 31 66 27 50 19 49 26
#> [326] 14 18 51 87  6 16 49  7 46  6 78 44 45 52 66 87 26 15 56 36 20 48 19 18 87
#> [351] 75  9  7 17 56 48 50 32 55 86 10 85  2 34 69 35 71 34 16 61 76 74 56 50 17
#> [376] 17 71 58 82 31  2 11 79 16 28 83  2 82 73 83 23 85 47 62 29 53 13 79  3 43
#> [401] 80 51 38 72 45 15 78  8 23 34 60 48 73 81 22 54 58 81 65  5 72 19 51  1 72
#> [426] 26 77 61 75 54 23 65 84 58 47 66 77 84  6 48 46 42 79 77 38  2 62 81 67 72
#> [451] 12  6  2 10 42 78 69 90 13 44 13 42 23 21 51 58 39 30 61  2  5 86 79 34 14
#> [476]  6 60 17 28 76 11 86 43 82 38 55 81 77 65 78 52 49 86  4 71 49  8 67 49  4