Simulate random birth times

This function simulates uniformly distributed birth times for a specified number of individuals from a provided vector

simulate_birth_times(N, times, age_min = 0)

Arguments

N

The number of individuals in the simulation

times

A vector of each time step in the simulation

age_min

A number matching the time resolution of times giving the youngest age possible by the end of the simulation; defaults to 0 which means individuals can be born up until the penultimate time step

Value

A vector of simulated birth times for each individual

See also

Examples

## Simulate random birth times for 500 individuals over 100 time steps and ensures 
## that all individuals are above 9 time steps old by the last time step
simulate_birth_times(500, 1:100, age_min=9) 
#>   [1] 44 69 46 28 75 22 37 35 86 17 77  9 44 32 74 40 88 18 69 17 56 28 83 21 29
#>  [26] 74  2 80  9 78 54 15  2 89 87 56 40 24 17 78 45 37 14 55  4 20 55 18 74 89
#>  [51] 63 63 72 54 70 12 72 21 76 28  1 60  3 12 24 16  7 67 83 72 21 27 35 78 64
#>  [76] 36 83 48 59 56 52 48 63 51 22 32 67 37 63 40 87 21 80 65 83  9 65 25 54 69
#> [101]  8 87 56 21 61  9 47 90 11 12 28 27 81 27 35 30 80 18 90 65 56 71 29 90 42
#> [126] 37 82 61 48 27 58 79 29 31  1 47 48 38 56  6 22 24  3  7 70 50 90 17 83 71
#> [151] 82 20  2 75 26 18  3 77 87  2  2 78 66 44 63 56 63 27 79 26 41 33  6 17 90
#> [176] 33 30 43 46 49 30 50 87 69 22 45 63 75 26 11  8 63 73 74 76 36 79 48 21 49
#> [201] 80 80 48 40 85 21 43 49 43 60 34 25 69 72 16 48 26 80 25 27 75 45 81 69  9
#> [226] 78 37 68 20 16  4 51 57 30 28 37 81 73 51 18 66  6 33 21 32 58 76 14 47 12
#> [251] 40 24 36 64 32 62 71 65 38 11 62 44 70 74 16 68 44 28 62  7 70 88 64 45 79
#> [276] 76 18 58 71 14 62 23 25 72 44 61 30 12 62  7 23 38 39  6 21 90 74  7 40 85
#> [301] 90 13 40 54 67 70 81 58 87 81  3 66 60 42 30  3 13 17 31 82 21 57 38 85 14
#> [326]  6 36 31 17 11 65 38 74 38 81 39 20 12 68 51 18 80 64 20 41 28 26 29 26 62
#> [351] 18 68 86 56 17 82 84 33 90 69 47 61 70 57 27 54 37 59 37  7 19 40 30 44 85
#> [376] 15 61 89 70 76 39 70  6 14 88 80 32 16 73 14 78 38 86 44 35 51 39 82  6 60
#> [401] 57 56 16  7 34 33 47 16 50 21 18  7 38 48 58 24 89 43 69 42 41 32 36 53 15
#> [426] 90 40 52 88 49 38 66 65 64 43 52 46  1 85  5 43 58 61 60 14 40 71 37 51 87
#> [451]  2 70 20 49 37 85  7 77 22 60 76 61 39 54 51 44 71 84 24 43  2 69 48 65  4
#> [476] 88 63 61 19 80 52 87 72 71 90 20 79 84 77  2 18 79 13  4 50 23 40 88 67 78