Simulate removal times for individuals

This function simulates uniformly distributed removal times for a specified number of individuals from the provided times vector. This might represent e.g., death or study attrition.

Usage

simulate_removal_times(
  N,
  times,
  birth_times,
  removal_min = 0,
  removal_max = max(times),
  prob_removal = 0
)

Arguments

N

The number of individuals in the simulation

times

A vector of each time step in the simulation

birth_times

A vector of birth times for each individual; defaults to NULL; if birth_times is not specified then the function will simulate uniformly distributed birth times for each individual from the times vector

removal_min

The minimum age at which an individual can be removed from the population. Defaults to 0

removal_max

The maximum age at which an individual can be removed from the population. Defaults to max(times)

prob_removal

The probability that an individual will be removed from the population during the simulation, representing e.g., death or study attrition. If set to NA, then removal time will be max(times)+1

Value

A vector of all individual's removal times is returned. NA represents no removal

See also

Examples

## Simulate random removal times for all individuals; Individuals have a 0.4 probability 
## of being removed at some time after they are 10 time steps old and before 
## they are 99 time steps old 
birth_times<-simulate_birth_times(500, 1:100, age_min=9) 
simulate_removal_times(500,1:100,birth_times, removal_min=10,removal_max=99, prob_removal=0.4)
#>   [1]  84  76 101 101 101 101  97  96  63 101  83 101 101 101 101 101 101  59
#>  [19] 101  73 101 101 101 101  98  92  88 101 101 101 101 101  70 101  86 101
#>  [37] 101  92 101  94 101  75  97 101  92  71 101 101  80 101 101 101  57  48
#>  [55] 101  83 101 101 101 101 101  89  78 101 101  44  39  64  70  79 101  38
#>  [73] 101  50 101  70 101 101 101  89 101 101 101  93  96 101  64  89 101 101
#>  [91]  86 101 101 101  68  89  38 101  81 101  55 101 101 101  49 101 101  83
#> [109] 101  55 101  86 101  95 101 101 101  48 101 101 101 101 101  70 101  92
#> [127]  84 101 101  95 101 101  95 101 101 101 101 101  98 101 101 101 101  97
#> [145] 101  81 101 101  75 101  49 101 101  85 101  75  97  58 101 101 101 101
#> [163] 101  75 101 101  94 101  67 101  97 101 101 101  51 101 101 101 101  74
#> [181] 101 101  73 101 101  87 101 101  97  84  41  65  75 101 101  43 101  44
#> [199] 101  96 101  99 101 101 101  83  81 101 101 101 101 101  69  65 101 101
#> [217]  39 101 101  12 101  57 101 101  90 101  97  91 101 101 101  89  95  98
#> [235]  87  94  54  22 101 101 101  45  77  95 101 101  72 101  90  73 101 101
#> [253]  84  65  87  31  27  81  43  90 101  10 101 101  78  91 101  67 101  18
#> [271]  87 101 101 101  90  41 101  42  17  97 101  57 101  69  49 101  99  73
#> [289] 101 101 101 101  86  92 101 101  88  32 101  69  89 101 101 101 101 101
#> [307] 101  63 101  94  90 101 101 101 101  53  18 101 101 101  92 101  79  84
#> [325]  27  77  92 101 101 101 101 101 101  85  95 101  79 101  28 101 101  86
#> [343] 101  99 101  53  37 101 101  82  95 101 101  88  45  81 101 101 101 101
#> [361] 101 101  92 101  53 101  68 101 101  69 101 101 101 101  62 101  96 101
#> [379] 101 101  39 101  66  92 101  92  79 101  59  59 101 101 101  37 101  56
#> [397] 101  98  89  83 101  76  95  86 101 101  32  84  52 101 101  66  48 101
#> [415]  96  80  86  45 101 101  95 101  60 101 101  89 101  80  88 101  41 101
#> [433] 101  97 101 101  80 101 101 101 101 101 101  84 101 101 101  95 101 101
#> [451] 101  73 101 101 101  50 101 101 101  44 101  48 101 101 101 101  81 101
#> [469] 101  95 101 101  56 101 101  39 101 101 101  42 101  70 101 101 101 101
#> [487] 101  44 101  86 101  79 101 101 101 101 101  52 101 101