Simulate removal times for individuals

This function simulates uniformly distributed removal times for a specified number of individuals from the provided times vector. This might represent e.g., death or study attrition.

Usage

simulate_removal_times(
  N,
  times,
  birth_times,
  removal_min = 0,
  removal_max = max(times),
  prob_removal = 0
)

Arguments

N

The number of individuals in the simulation

times

A vector of each time step in the simulation

birth_times

A vector of birth times for each individual; defaults to NULL; if birth_times is not specified then the function will simulate uniformly distributed birth times for each individual from the times vector

removal_min

The minimum age at which an individual can be removed from the population. Defaults to 0

removal_max

The maximum age at which an individual can be removed from the population. Defaults to max(times)

prob_removal

The probability that an individual will be removed from the population during the simulation, representing e.g., death or study attrition. If set to NA, then removal time will be max(times)+1

Value

A vector of all individual's removal times is returned. NA represents no removal

See also

Examples

## Simulate random removal times for all individuals; Individuals have a 0.4 probability 
## of being removed at some time after they are 10 time steps old and before 
## they are 99 time steps old 
birth_times<-simulate_birth_times(500, 1:100, age_min=9) 
simulate_removal_times(500,1:100,birth_times, removal_min=10,removal_max=99, prob_removal=0.4)
#>   [1] 101 101 101  78 101 101  90 101 101  88  86 101 101  88 101 101 101  93
#>  [19]  39 101  31 101 101 101 101  97 101  97  88 101  23  51  81  88  73 101
#>  [37] 101  82 101  96 101  86  77  92 101  93 101 101 101  99 101 101  84  82
#>  [55] 101 101  91 101 101 101 101 101 101 101  82 101  89 101  74 101 101 101
#>  [73]  83  79 101 101 101 101  51 101  81 101 101 101  88 101 101 101 101  94
#>  [91] 101  73  68  69  83 101  80 101 101  83 101  41 101  93  97 101 101 101
#> [109]  74 101  85 101 101 101 101 101 101  81 101  78  66 101 101  58 101  53
#> [127] 101 101 101 101  45 101  99  66 101  76 101  97 101  55  75 101  94 101
#> [145] 101 101  96  96 101  93  77 101  58 101 101 101 101 101  92 101  47  37
#> [163] 101 101  83 101 101 101  53 101  88 101 101 101  53 101 101 101  42 101
#> [181]  37  99  90 101  88 101  27 101 101 101 101 101 101 101 101  95  76 101
#> [199]  95  97 101  64  73  99  93 101 101  92 101  31 101 101  96 101 101  36
#> [217]  78  83 101  26  46  67 101 101 101 101 101 101 101  48  92 101  63  85
#> [235] 101 101  99  66 101  82 101 101  84  84  57 101 101 101  65 101  72 101
#> [253] 101 101 101 101  37  85  92 101  90 101  45  68 101 101 101 101 101  64
#> [271] 101 101  75 101  84  92 101 101 101 101 101  60  94 101 101 101  79  89
#> [289] 101  63 101  67 101 101 101 101 101 101  97 101  95  96  85 101 101  87
#> [307] 101 101  59 101 101  56 101 101 101 101 101  30 101  38 101  94 101 101
#> [325]  96 101  46 101 101 101 101 101 101 101 101 101 101 101 101  87  95 101
#> [343]  79  99 101  55 101 101  14 101 101  66  94 101  79 101 101 101  95  76
#> [361]  76  92 101  73 101  92 101 101  96  47 101 101 101  99  63 101 101  45
#> [379]  41 101 101 101 101 101 101  99 101 101 101  95 101 101 101 101  68  48
#> [397] 101  64 101 101 101 101  66  57  85  91 101  70 101 101 101  96 101  52
#> [415]  99  78 101  79 101 101 101  85 101 101 101  87 101 101  55 101  96 101
#> [433]  93 101 101  51  51 101  95  58  89  59 101 101  53 101  72 101  83  76
#> [451]  68  29  91 101  56 101  79 101  90  27 101  68 101 101 101 101  89 101
#> [469]  84 101  96 101 101  53 101 101 101 101 101  50  83 101 101  80 101  85
#> [487] 101 101  82 101 101 101  96  85 101  87 101 101 101 101