Simulate removal times for individuals

This function simulates uniformly distributed removal times for a specified number of individuals from the provided times vector. This might represent e.g., death or study attrition.

simulate_removal_times(
  N,
  times,
  birth_times,
  removal_min = 0,
  removal_max = max(times),
  prob_removal = 0
)

Arguments

N

The number of individuals in the simulation

times

A vector of each time step in the simulation

birth_times

A vector of birth times for each individual; defaults to NULL; if birth_times is not specified then the function will simulate uniformly distributed birth times for each individual from the times vector

removal_min

The minimum age at which an individual can be removed from the population. Defaults to 0

removal_max

The maximum age at which an individual can be removed from the population. Defaults to max(times)

prob_removal

The probability that an individual will be removed from the population during the simulation, representing e.g., death or study attrition. If set to NA, then removal time will be max(times)+1

Value

A vector of all individual's removal times is returned. NA represents no removal

See also

Examples

## Simulate random removal times for all individuals; Individuals have a 0.4 probability 
## of being removed at some time after they are 10 time steps old and before 
## they are 99 time steps old 
birth_times<-simulate_birth_times(500, 1:100, age_min=9) 
simulate_removal_times(500,1:100,birth_times, removal_min=10,removal_max=99, prob_removal=0.4)
#>   [1] 101 101 101 101 101 101 101 101 101 101 101 101  53  90 101 101 101 101
#>  [19] 101  85  87 101 101 101  95  96  88  85  99 101 101 101  75  98  74  82
#>  [37] 101  57  49  56 101  91  89 101 101 101  79 101 101 101  39  88  87 101
#>  [55]  90 101 101 101 101 101 101 101 101  39 101 101  90 101  70 101  26 101
#>  [73]  49 101  98 101 101 101 101  89 101  81 101 101 101  95 101 101  65 101
#>  [91] 101 101 101 101 101  92 101 101  87 101  75 101 101  99 101 101 101 101
#> [109]  90 101  59  81  94 101  61 101 101 101 101  99 101 101  33 101 101 101
#> [127] 101 101  59 101 101 101  74 101 101  68 101 101  79  98  99 101  71  56
#> [145] 101 101 101 101 101 101 101 101  94  68  83  92  72 101 101  74  92 101
#> [163] 101 101 101  95  67 101  96  82 101 101 101 101 101 101  37 101  94 101
#> [181] 101 101 101  93 101  82 101 101 101 101 101  78 101  38 101 101  38 101
#> [199] 101  95 101 101  84  81  83 101  10 101 101 101 101 101  96 101 101 101
#> [217]  82 101 101  76 101 101 101  50  75 101  88  95  99 101  94 101 101  93
#> [235] 101  86  23 101 101 101  91  93  28  55  43 101 101 101  77  73 101  74
#> [253]  79 101 101  95 101  66  86 101 101 101 101 101 101  44 101 101 101 101
#> [271]  97  90  58 101 101 101 101 101 101 101 101 101 101  65 101 101  46  40
#> [289] 101 101 101  63  87 101 101 101 101 101  91  94  73 101 101 101 101  34
#> [307] 101  94 101 101  93  73  95  90  13 101  60 101  42  72 101  67  50 101
#> [325]  52  91  80 101 101 101  30 101 101  92  76 101  48 101 101 101  43  57
#> [343] 101  64 101 101 101 101  97 101  42 101 101 101  86 101  96  97 101  65
#> [361]  74  83 101 101 101 101 101 101  96 101 101  79 101  69 101  64 101  96
#> [379] 101 101  81  96 101 101  88  98  69  71 101 101 101 101 101  40 101  84
#> [397] 101  73 101 101  91  51  92 101  75  75 101  94 101 101 101  86  61 101
#> [415] 101 101  87 101 101 101 101  81  65  88 101 101  92  69  89 101  51 101
#> [433]  97  76  64  61 101 101  94  60 101 101 101  74  75 101  52 101 101  43
#> [451] 101 101  86 101  92  47  84 101  61 101 101  76  12  89  91 101  64 101
#> [469] 101 101  81  96  22 101  97  85  92  93  64  12  89 101 101  97  36 101
#> [487] 101 101  77 101 101 101  16 101 101 101  89 101 101  92