Simulate removal times for individuals

This function simulates uniformly distributed removal times for a specified number of individuals from the provided times vector. This might represent e.g., death or study attrition.

simulate_removal_times(
  N,
  times,
  birth_times,
  removal_min = 0,
  removal_max = max(times),
  prob_removal = 0
)

Arguments

N
The number of individuals in the simulation

times
A vector of each time step in the simulation

birth_times
A vector of birth times for each individual; defaults to NULL; if birth_times is not specified then the function will simulate uniformly distributed birth times for each individual from the times vector

removal_min
The minimum age at which an individual can be removed from the population. Defaults to 0

removal_max
The maximum age at which an individual can be removed from the population. Defaults to max(times)

prob_removal
The probability that an individual will be removed from the population during the simulation, representing e.g., death or study attrition. If set to NA, then removal time will be max(times)+1

Value

A vector of all individual's removal times is returned. NA represents no removal

See also

Examples

## Simulate random removal times for all individuals; Individuals have a 0.4 probability 
## of being removed at some time after they are 10 time steps old and before 
## they are 99 time steps old 
birth_times<-simulate_birth_times(500, 1:100, age_min=9) 
simulate_removal_times(500,1:100,birth_times, removal_min=10,removal_max=99, prob_removal=0.4)
#>   [1] 101  74  98 101  46 101 101  89 101  83  59 101 101  92  74 101 101  79
#>  [19]  29 101  33 101 101 101 101  31 101 101  15 101  94  79 101 101 101  64
#>  [37]  63 101  78  84 101 101 101 101  85  71  32 101 101  91  92 101 101 101
#>  [55] 101 101 101 101 101 101  98  50 101  59 101  96 101 101 101 101  80 101
#>  [73] 101 101 101 101  52 101  80  73  87  85  91  89 101  98  95 101 101 101
#>  [91]  72 101  87 101 101  18 101 101  86 101 101  95  26 101 101  87  96  90
#> [109] 101  92  74  97 101 101 101  86 101 101 101  91 101  60  48 101  63 101
#> [127] 101 101  76 101 101  72  26 101  79 101  87  69  87 101 101 101 101  98
#> [145]  79 101  87  86 101 101  19 101  93  83 101 101  48  69  23  84 101 101
#> [163] 101 101 101 101  89 101 101 101  78 101 101  31 101 101 101  42  99 101
#> [181] 101 101 101 101 101  86 101  33  97  58  67 101 101 101  88  68 101  90
#> [199]  38  42 101 101  45  92 101 101  89 101 101 101  85 101  95 101  62  67
#> [217]  88 101  70 101  83 101  96 101  95 101  58  97 101  98 101 101 101 101
#> [235] 101 101 101 101 101 101 101 101 101  76  71 101  78  88 101  84 101  70
#> [253] 101  89 101 101 101 101  76 101  84 101 101 101 101  64  48 101 101 101
#> [271]  54 101  64 101 101  74 101  39  19 101 101 101 101  62  78  98  82  92
#> [289] 101 101 101  41  43 101 101 101 101 101 101  97 101  44 101 101 101 101
#> [307]  91 101 101  53  60  99 101 101 101  85  80 101 101  34 101  98 101  56
#> [325] 101  60  87  98  87  65  77 101 101 101  88 101 101 101  84 101 101  99
#> [343]  79  87 101  44 101  52 101  93  86  93  67 101  54  62 101 101  96  94
#> [361] 101  82  95  87 101  51 101 101  47  91  71 101 101 101  93  84 101 101
#> [379]  93 101  35  53 101  78 101  78 101 101 101  92  58  17  63  50 101 101
#> [397] 101 101 101 101 101 101 101  55  89 101  78 101 101  74  52 101  73 101
#> [415]  98 101  89 101 101  91 101  38  58 101  61 101 101 101 101 101  29  52
#> [433] 101  42 101 101 101 101  96  59 101  99  87  91 101  74 101 101 101  90
#> [451] 101  94 101 101  64  15 101  75 101 101 101  94 101 101 101  79 101 101
#> [469]  78 101  95 101  72 101 101  14  81  95 101 101  45  96  69 101 101 101
#> [487] 101 101  46  67 101 101  94 101  67  88  52 101 101 101